feat: Managing books in the bookshelf and adding new books from search view
Created new components: Book, BookShelf, ListMyReads, ListSearchBooks.
Managing books on different shelves is done, it is also possible to add
new books from search view. But books in the search view are not yet in sync
with the books that are already in the bookshelf. In addition, search query
is buggy and slow.